backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
yt-dlp>=2024.03.10  
groq==0.9.0
langchain-core==0.2.43
langchain-community==0.2.19
langchain-groq==0.1.10
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.4
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.2
pinecone-client==5.0.1
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.32.3
tiktoken==0.7.0
huggingface-hub==0.23.0
httpx==0.27.2
pip-system-certs